[#782] fix voting page header calendar
